App과 MainPage 분리
