+++ docs/payment/요구사항 분석 및 정리.md
@@ -10,6 +10,9 @@ React, Typescript, TailwindCSS
 ### 문서 작성자: 
 민진홍
 
+### 디자인: 
+https://www.figma.com/design/DKeR5wR65RkxcAWReoBccl/%EC%9D%91%EC%9A%A9%EC%86%8C%ED%94%84%ED%8A%B8%EC%9B%A8%EC%96%B4-%EB%94%94%EC%9E%90%EC%9D%B8-%EC%8B%9C%EC%95%88?node-id=1-130&node-type=canvas&t=Vm7MVuvf92oyiqWD-0
+
 ## 📝 고객 요구사항 정리
 - 신용 카드 등록
   - 민감 정보 숨김 처리
