+++ project/src/pages/PaymentPage.tsx
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+import Cart from '../components/cart';
+import Product from '../components/product';
+
+function PaymentPage() {
+  return(
+    <>
+    </>
+  )
+}
+
+export default PaymentPage;
